backend/requirements.txt
# backend
Flask==3.0.3
flask-cors==4.0.1
pandas==2.2.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.0
Flask-Caching==2.3.0

# allocation_algorithm
PuLP==2.8.0